test(sim): carry replica ids as u16, split voting vs total membership
The simulator distinguishes the VOTING count (drives quorum expectations, the
minority fault budget, and the crash/isolate/asym victim pickers) from the
total `node_count` (sizes every per-replica vector, the routing target space,
storage seeding, and the fan-out loops). Every replica-id carrier widens to
u16 with checked `usize` indexing — `Target::Replica`, the partition/one-way
APIs and their 2D vectors, the pending-view probe, the client/protocol
scheduling, the recipient reconstruction, the donor accessor, and the VOPR
chaos casts — so a member id beyond the u8 range cannot truncate or alias.

The minority budget and the budget-charged victim pickers range over the
voting set only, so a non-voting member's outage can never substitute for a
voter's against the quorum survivability assumption; the per-member fault and
routing vectors span the full membership. A new measured gate records the
maximum live checkpoint envelope across the digest corpus and asserts it stays
well below the unchunked-vs-chunked cap, so the carrier-overhead change cannot
have flipped any sync decision. A routing regression constructs a member id
above 255 and proves it is reached without aliasing a low id.

With zero non-voting members the voting count equals node_count, so the full
VOPR digest is byte-identical to the pre-change sweep across 64 seeds.

+++ tests/viewstamp-simulation/src/network.rs
@@ -3,8 +3,10 @@ use viewstamp_proto::{Instant, Message, Peer};
 /// Where a delivered message goes.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Target {
-  /// A replica, by index.
-  Replica(u8),
+  /// A replica, by index. The index spans the full membership (`0..node_count`), so it is a
+  /// `u16` to match [`ReplicaId::get`](viewstamp_proto::ReplicaId::get) — a high member id cannot
+  /// truncate or alias a low one.
+  Replica(u16),
   /// A client, by id.
   Client(u128),
 }
